+++ src/redturtle/prenotazioni/testing.py
@@ -3,6 +3,7 @@ from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_FIXTURE
 from plone.app.testing import applyProfile
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
+from plone.app.testing import MOCK_MAILHOST_FIXTURE
 from plone.app.testing import PloneSandboxLayer
 from plone.restapi.testing import PloneRestApiDXLayer
 from plone.testing import z2
@@ -22,7 +23,7 @@ except ImportError:
 
 
 class RedturtlePrenotazioniLayer(PloneSandboxLayer):
-    defaultBases = (PLONE_APP_CONTENTTYPES_FIXTURE,)
+    defaultBases = (PLONE_APP_CONTENTTYPES_FIXTURE, MOCK_MAILHOST_FIXTURE)
 
     def setUpZope(self, app, configurationContext):
         # Load any other ZCML that is required for your tests.